Add option to use a classpath jar (#95, #152)
On Windows the classpath passed to the compiler and server runner are
problematic as they often exceed the command line limit provided by
Windows. To circumvent that we use a classpath jar that defines the
project classpath instead in the manifest of the Jar.

This feature should be considered experimental and is off by default.