Update JettyServerRunner.java
Setting ParentLoaderPriority to true seem to cause classes in dependency jar to be loaded by the system class loader and not by the webapp one. Causing conflict error when try to perform reflection operations with web app loaded classes.
If you set it to false (the default), Jetty uses standard webapp classloading priority.